Refactor core module (#13)

+++ src/main/java/com/raunlo/checklist/core/entity/ChangeOrderRequest.java
@@ -10,7 +10,7 @@ import lombok.With;
 @AllArgsConstructor
 @NoArgsConstructor
 public class ChangeOrderRequest {
-    private long taskId;
-    private long newOrderNumber;
+    private int taskId;
+    private int newOrderNumber;
     private long checklistId;
 }
